menu de escolhas em main.py
