+++ dados_login.json
@@ -4,5 +4,17 @@
     },
     "gustavo": {
         "senha": "3456"
+    },
+    "reginaldo": {
+        "senha": "235343"
+    },
+    "amelia": {
+        "senha": "98876"
+    },
+    "Juvenal da Silva": {
+        "senha": "2345"
+    },
+    "Mariana Domingues": {
+        "senha": "1234"
     }
 }
